refactor(lib): Extract duplicate deepMerge() to lib/core/deep-merge.ts
- Create lib/core/deep-merge.ts with type-safe generic implementation
- Remove duplicate from lib/content/metadata-schemas.ts (line 342)
- Remove duplicate from lib/settings/utils.ts (line 98)
- Update both files to import from shared utility
- Benefits: DRY principle, single source of truth, better type safety
