+++ lib/content/metadata-schemas.ts
@@ -24,6 +24,8 @@
  * ```
  */
 
+import { deepMerge } from "@/lib/core/deep-merge";
+
 // ============================================================
 // External Provider Integrations
 // ============================================================
@@ -334,33 +336,6 @@ export function setMetadata(
 ): FileMetadata {
   const current = getMetadata(existing);
   return deepMerge(current, updates);
-}
-
-/**
- * Deep merge two objects
- */
-function deepMerge<T>(target: T, source: Partial<T>): T {
-  const result = { ...target };
-
-  for (const key in source) {
-    const sourceValue = source[key];
-    const targetValue = result[key];
-
-    if (
-      sourceValue &&
-      typeof sourceValue === "object" &&
-      !Array.isArray(sourceValue) &&
-      targetValue &&
-      typeof targetValue === "object" &&
-      !Array.isArray(targetValue)
-    ) {
-      result[key] = deepMerge(targetValue, sourceValue) as T[Extract<keyof T, string>];
-    } else if (sourceValue !== undefined) {
-      result[key] = sourceValue as T[Extract<keyof T, string>];
-    }
-  }
-
-  return result;
 }
 
 /**
